Move OpenAPI Versions to fields and the object to the OAS spec
