+++ amati/fields/openapi_versions.py
@@ -1,6 +1,5 @@
 """
-Validates the OpenAPI Object - §4.8.1:
-
+Validates the OpenAPI Version
 """
 
 from typing import Annotated
@@ -8,15 +7,12 @@ from typing import Annotated
 from pydantic import AfterValidator, Field
 
 from amati.logging import Log, LogMixin
-from amati.validators import title
-from amati.validators.generic import GenericObject
-from amati.validators.oas311 import InfoObject
 from amati.validators.reference_object import Reference, ReferenceModel
 
 reference: Reference = ReferenceModel(
-    title=title,
-    url='https://spec.openapis.org/oas/latest.html#openapi-object',
-    section='OpenAPI Object'
+    title='OpenAPI Initiative Publications',
+    url='https://spec.openapis.org/#openapi-specification',
+    section='OpenAPI Specification '
 )
 
 OPENAPI_VERSIONS = ['1.0', '1.1', '1.2', '2.0', '3.0', '3.0.1', '3.0.2', '3.0.3', '3.1', '3.1.1']
@@ -24,7 +20,7 @@ OPENAPI_VERSIONS = ['1.0', '1.1', '1.2', '2.0', '3.0', '3.0.1', '3.0.2', '3.0.3'
 
 def _validate_after_openapi(value: str) -> str:
     if value not in OPENAPI_VERSIONS:
-        LogMixin.log(Log(f"{value} is not a valid OpenAPI version.", ValueError))
+        LogMixin.log(Log(f"{value} is not a valid OpenAPI version.", ValueError, reference))
     return value
 
 OpenAPI = Annotated[
@@ -32,9 +28,3 @@ OpenAPI = Annotated[
     Field(strict=True),
     AfterValidator(_validate_after_openapi)
     ]
-
-
-class OpenAPIObject(GenericObject):
-    openapi: OpenAPI
-    info: InfoObject
-    _reference: Reference = reference # type: ignore
